pglbox version 3 and 4 into develop, support training big language model for ps mode; support multi-node sample and train in graph-based work.

+++ paddle/phi/kernels/xpu/argsort_kernel.cc
@@ -182,7 +182,7 @@ void ArgsortKernel(const Context& dev_ctx,
 
   if (rank == 0) {
     phi::Copy<Context>(dev_ctx, input, dev_ctx.GetPlace(), false, output);
-    phi::funcs::set_constant(dev_ctx, indices, 0);
+    phi::funcs::set_constant(dev_ctx, indices, static_cast<int64_t>(0));
     return;
   }
